テスト用のプログラムをテストファイルに移動
